ljwx-bigscreen/bigscreen/requirements-docker.txt
redis>=4.5.0
flask>=2.3.0
flask-sqlalchemy>=3.0.0
pymysql>=1.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
requests>=2.31.0
numpy>=1.24.0,<2.0.0
cryptography>=41.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
pytz>=2023.3
qrcode>=7.4.0
pandas>=2.0.0,<3.0.0
prometheus-client>=0.17.0
schedule>=1.2.0 
